nptl: add pthread_cond_clockwait GNU extension
As defined by POSIX.1-2001 the 'abstime' timeout of
pthread_cond_timedwait() shall be based on CLOCK_REALTIME. The added
pthread_cond_clockwait() function behaves the same except that it lets
the user specify the clock against which the timeout is measured. The
implementation is based on [1] which covers all functions that take an
absolute struct timespec timeout.

[1] http://austingroupbugs.net/view.php?id=1216

Signed-off-by: Jan Klötzke <jan@kloetzke.net>