Semaphores: add sem_clockwait GNU extension [BZ #14717]
As defined by POSIX.1-2001 the 'abstime' timeout of  sem_timedwait()
shall be based on CLOCK_REALTIME. This has the unpleasant effect that
application timeouts are affected by system clock adjustments.

The added sem_clockwait() function behaves just like sem_timedwait() but
lets the user specify the clock against which the timeout is measured.
The implementation is based on [1] which covers all functions that take an
absolute struct timespec timeout.

[1] http://austingroupbugs.net/view.php?id=1216

Signed-off-by: Jan Klötzke <jan@kloetzke.net>